More game functionalities added

+++ src/main/java/nl/cesar/tictactoe/repository/PlayerRepository.java
@@ -8,9 +8,9 @@ import org.springframework.stereotype.Repository;
 import nl.cesar.tictactoe.domain.Player;
 
 @Repository
-public interface PlayerRepository extends CrudRepository<Player, Integer>{
+public interface PlayerRepository extends CrudRepository<Player, Long>{
 
 	Optional<Player> findByUsername(String userName);
-	Optional<Player> findById(Integer id);
+	Optional<Player> findById(Long id);
 
 }
